refactor dataloader and fi dimension issue
