+++ tests/unit_tests/test_flux_dataloader.py
@@ -4,17 +4,16 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-
 import torch
 from torchtitan.config_manager import JobConfig
-from torchtitan.datasets.flux_datasets import build_flux_dataloader
+from torchtitan.datasets.flux_dataset import build_flux_dataloader
 from torchtitan.experiments.flux.utils import load_clip, load_t5
 
 
 class TestFLUXDataLoader:
     def test_FLUX_dataloader(self):
         dataset_name = "cc12m"
-        batch_size = 1  # batch_size = 4 will cause OOM
+        batch_size = 2  # batch_size = 4 will cause OOM
         seq_len = 512
         world_size = 4
         rank = 0
@@ -27,10 +26,11 @@ class TestFLUXDataLoader:
 
         for k, v in input_data.items():
             print(k, v.shape)
-            input_data[k] = v.cpu()
 
-        assert len(input_data) == 6
-        # TODO(jianiw): Add more data shape check once we finalize input and model
+        assert len(input_data) == 3  # (image, clip_tokens, t5_tokens)
+        assert input_data["image"].shape == (batch_size, 3, 256, 256)
+        assert input_data["clip_tokens"].shape[0] == batch_size
+        assert input_data["t5_tokens"].shape == (batch_size, 1, 512, 512)
 
     def _build_dataloader(
         self, dataset_name, batch_size, seq_len, world_size, rank, seed, device="cpu"
@@ -52,9 +52,8 @@ class TestFLUXDataLoader:
         return build_flux_dataloader(
             dp_world_size=world_size,
             dp_rank=rank,
-            t5_encoder=load_t5(device=device, max_length=seq_len),
-            clip_encoder=load_clip(device=device),
+            t5_encoder=load_t5(max_length=512).to(device),
+            clip_encoder=load_clip().to(device),
             job_config=config,
             infinite=False,
-            device=device,
         )
